+++ example/lib/main.dart
@@ -1,11 +1,13 @@
-import 'package:flutter/material.dart';
 import 'package:adaptive_spinner/adaptive_spinner.dart';
+import 'package:flutter/material.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -14,12 +16,14 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: MyHomePage(),
+      home: const MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
+  const MyHomePage({Key? key}) : super(key: key);
+
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
@@ -28,12 +32,12 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('AdaptiveSpinner Demo')),
+      appBar: AppBar(title: const Text('AdaptiveSpinner Demo')),
       body: Padding(
         padding: const EdgeInsets.all(20),
         child: Center(
-          child: Column(
-            children: [
+          child: ListView(
+            children: const [
               AdaptiveSpinner(),
               Divider(),
               AdaptiveSpinner(
@@ -45,8 +49,9 @@ class _MyHomePageState extends State<MyHomePage> {
                 withContainer: true,
                 width: double.infinity,
                 height: 100,
-                withMessage: 'With container\nand with\nright-aligned text\n(width: double.infinity, height: 100)',
-                color: Colors.black12,
+                withMessage:
+                    'With container\nand with\nright-aligned text\n(width: double.infinity, height: 100)',
+                withContainerColor: Colors.black12,
                 withTextAlign: TextAlign.right,
               ),
               Divider(),
@@ -70,8 +75,16 @@ class _MyHomePageState extends State<MyHomePage> {
                   fontSize: 15,
                   fontWeight: FontWeight.bold,
                 ),
-                withRadius: 50,
-                withMessage: 'With radius (50)',
+                withRadius: 20.0,
+                withMessage: 'With radius (20)',
+              ),
+              Divider(),
+              AdaptiveSpinner(
+                withContainerColor: Colors.pink,
+                withSpinnerColor: Colors.greenAccent,
+                withDuration: const Duration(milliseconds: 500),
+                withMessage:
+                    'With colored spinner and container\nand fast spinning (Duration: 500ms)',
               ),
             ],
           ),