+++ CHANGELOG.md
@@ -1,8 +1,20 @@
+## [3.0.0] - BREAKING - SpinKit
+* AdaptiveSpinner now uses SpinKit instead of Flutter's default activity indicators
+* Removed the `centeredContainer` property
+* Changed the `color` property, now is `withContainerColor`
+
 ## [2.0.0] - BREAKING - Null-safety support
+
 ## [1.1.1] - Bugfix spinner crashing if no radius was specified
+
 ## [1.1.0] - The spinner size is now customizable
+
 ## [1.0.0] - BREAKING - Now the container is fully customizable
+
 ## [0.1.3] - Formatted code
+
 ## [0.1.2] - Formatted code
+
 ## [0.1.1] - Improved Dartdoc
+
 ## [0.1.0] - Initial release